+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - Removed `name` parameter from `init_project` function and `management` API ([#1227](https://github.com/neptune-ai/neptune-client/pull/1227))
 - Monitoring namespace based on hostname, process id and thread id ([#1222](https://github.com/neptune-ai/neptune-client/pull/1222))
 - Removed deprecated `--run` option from `neptune sync` command ([#1231](https://github.com/neptune-ai/neptune-client/pull/1231))
+- Removed `Run._short_id` property ([#1234](https://github.com/neptune-ai/neptune-client/pull/1234))
+- Removed `get_run_url` method ([#1238](https://github.com/neptune-ai/neptune-client/pull/1238))
+- Removed `neptune.new.sync` module ([#1240](https://github.com/neptune-ai/neptune-client/pull/1240))
 - Added exception for unsupported types ([#1229](https://github.com/neptune-ai/neptune-client/pull/1229))
 
 ## neptune-client 0.16.18